Added function get_enabled to export vring enable status.
The vring is disabled and RX/TX queue is set to
not-ready during snapshot. But vhost-device process
keeps access vring and queue after snapshot which
caused system hang. We need to export vring enabled
and disabled status for snapshot feature so that
the vhost-device would skip vring/queue access
aftr snapshot to avoid hang.

Signed-off-by: Wei-Chung Hsu <weihsu@google.com>